fix: Remove iox_object_store crate; move ParquetFilePath to parquet_file
