--- iox_object_store/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "iox_object_store"
-version = "0.1.0"
-edition = "2021"
-description = "IOx-specific semantics wrapping the general-purpose object store crate"
-
-[dependencies]
-bytes = "1.0"
-data_types = { path = "../data_types" }
-data_types2 = { path = "../data_types2" }
-futures = "0.3"
-object_store = { path = "../object_store" }
-observability_deps = { path = "../observability_deps" }
-snafu = "0.7"
-tokio = { version = "1.18", features = ["macros", "parking_lot", "sync", "rt"] }
-tokio-stream = "0.1"
-uuid = { version = "0.8", features = ["v4"] }
-workspace-hack = { path = "../workspace-hack"}
-
-[dev-dependencies] # In alphabetical order
-test_helpers = { path = "../test_helpers" }
