--- iox_object_store/src/paths/transaction_file.rs
@@ -1,384 +1,0 @@
-use object_store::{
-    path::{parsed::DirsAndFileName, ObjectStorePath, Path as ObjStoPath},
-    Result,
-};
-use snafu::{ensure, OptionExt, ResultExt, Snafu};
-use std::str::FromStr;
-use uuid::Uuid;
-
-/// File suffix for transaction files in object store.
-const TRANSACTION_FILE_SUFFIX: &str = "txn";
-
-/// File suffix for checkpoint files in object store.
-const CHECKPOINT_FILE_SUFFIX: &str = "ckpt";
-
-/// Location of a catalog transaction file within a database's object store.
-/// The exact format is an implementation detail and is subject to change.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct TransactionFilePath {
-    /// Transaction revision
-    pub revision_counter: u64,
-    /// Transaction identifier
-    pub uuid: Uuid,
-    suffix: TransactionFileSuffix,
-}
-
-impl TransactionFilePath {
-    /// Create a new file path to store transaction info.
-    pub fn new_transaction(revision_counter: u64, uuid: Uuid) -> Self {
-        Self {
-            revision_counter,
-            uuid,
-            suffix: TransactionFileSuffix::Transaction,
-        }
-    }
-
-    /// Create a new file path to store checkpoint info.
-    pub fn new_checkpoint(revision_counter: u64, uuid: Uuid) -> Self {
-        Self {
-            revision_counter,
-            uuid,
-            suffix: TransactionFileSuffix::Checkpoint,
-        }
-    }
-
-    /// Returns true if this path is to a checkpoint file; false otherwise.
-    pub fn is_checkpoint(&self) -> bool {
-        self.suffix == TransactionFileSuffix::Checkpoint
-    }
-
-    /// Turn this into directories and file names to be added to a root path
-    pub fn relative_dirs_and_file_name(&self) -> DirsAndFileName {
-        let mut result = DirsAndFileName::default();
-
-        // pad number: `u64::MAX.to_string().len()` is 20
-        result.push_dir(format!("{:0>20}", self.revision_counter));
-
-        let file_name = format!("{}.{}", self.uuid, self.suffix.as_str());
-        result.set_file_name(file_name);
-        result
-    }
-
-    /// Create from serialized protobuf strings.
-    pub fn from_relative_dirs_and_file_name(
-        dirs_and_file_name: &DirsAndFileName,
-    ) -> Result<Self, TransactionFilePathParseError> {
-        let mut directories = dirs_and_file_name.directories.iter();
-
-        let revision_counter = directories
-            .next()
-            .context(MissingRevisionCounterSnafu)?
-            .to_string()
-            .parse()
-            .context(InvalidRevisionCounterSnafu)?;
-
-        ensure!(directories.next().is_none(), UnexpectedDirectorySnafu);
-
-        let file_name = dirs_and_file_name
-            .file_name
-            .as_ref()
-            .context(MissingFileNameSnafu)?
-            .to_string();
-        let mut parts = file_name.split('.');
-
-        let uuid = parts
-            .next()
-            .context(MissingUuidSnafu)?
-            .parse()
-            .context(InvalidUuidSnafu)?;
-
-        let suffix = parts
-            .next()
-            .context(MissingSuffixSnafu)?
-            .parse()
-            .context(InvalidSuffixSnafu)?;
-
-        ensure!(parts.next().is_none(), UnexpectedExtensionSnafu);
-
-        Ok(Self {
-            revision_counter,
-            uuid,
-            suffix,
-        })
-    }
-
-    // Deliberately pub(crate); this transformation should only happen within this crate
-    pub(crate) fn from_absolute(
-        absolute_path: ObjStoPath,
-    ) -> Result<Self, TransactionFilePathParseError> {
-        let absolute_path: DirsAndFileName = absolute_path.into();
-
-        let mut absolute_dirs = absolute_path.directories.into_iter().fuse();
-
-        // The number of `next`s here needs to match the total number of directories in
-        // iox_object_store transactions_path
-        absolute_dirs.next(); // "dbs"
-        absolute_dirs.next(); // database uuid
-        absolute_dirs.next(); // "transactions"
-
-        let remaining = DirsAndFileName {
-            directories: absolute_dirs.collect(),
-            file_name: absolute_path.file_name,
-        };
-
-        Self::from_relative_dirs_and_file_name(&remaining)
-    }
-}
-
-#[derive(Snafu, Debug, PartialEq)]
-#[allow(missing_docs)]
-pub enum TransactionFilePathParseError {
-    #[snafu(display("Could not find required revision counter"))]
-    MissingRevisionCounter,
-
-    #[snafu(display("Could not parse revision counter: {}", source))]
-    InvalidRevisionCounter { source: std::num::ParseIntError },
-
-    #[snafu(display("Too many directories found"))]
-    UnexpectedDirectory,
-
-    #[snafu(display("Could not find required file name"))]
-    MissingFileName,
-
-    #[snafu(display("Could not find required UUID"))]
-    MissingUuid,
-
-    #[snafu(display("Could not parse UUID: {}", source))]
-    InvalidUuid { source: uuid::Error },
-
-    #[snafu(display("Could not find required suffix"))]
-    MissingSuffix,
-
-    #[snafu(display("Invalid suffix: {}", source))]
-    InvalidSuffix {
-        source: TransactionFileSuffixParseError,
-    },
-
-    #[snafu(display("Too many extensions found"))]
-    UnexpectedExtension,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-enum TransactionFileSuffix {
-    Transaction,
-    Checkpoint,
-}
-
-impl TransactionFileSuffix {
-    fn as_str(&self) -> &'static str {
-        match self {
-            Self::Transaction => TRANSACTION_FILE_SUFFIX,
-            Self::Checkpoint => CHECKPOINT_FILE_SUFFIX,
-        }
-    }
-}
-
-#[derive(Snafu, Debug, PartialEq)]
-#[allow(missing_docs)]
-pub enum TransactionFileSuffixParseError {
-    #[snafu(display("Unknown suffix: {}", suffix))]
-    UnknownSuffix { suffix: String },
-}
-
-impl FromStr for TransactionFileSuffix {
-    type Err = TransactionFileSuffixParseError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            TRANSACTION_FILE_SUFFIX => Ok(Self::Transaction),
-            CHECKPOINT_FILE_SUFFIX => Ok(Self::Checkpoint),
-            suffix => UnknownSuffixSnafu { suffix }.fail(),
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{paths::ALL_DATABASES_DIRECTORY, IoxObjectStore, RootPath};
-    use object_store::{DynObjectStore, ObjectStoreImpl};
-    use std::sync::Arc;
-    use test_helpers::assert_error;
-
-    /// Creates a new in-memory object store. These tests rely on the `Path`s being of type
-    /// `DirsAndFileName` and thus using object_store::path::DELIMITER as the separator
-    fn make_object_store() -> Arc<DynObjectStore> {
-        Arc::new(ObjectStoreImpl::new_in_memory())
-    }
-
-    #[test]
-    fn is_checkpoint_works() {
-        let uuid = Uuid::new_v4();
-
-        let transaction = TransactionFilePath::new_transaction(0, uuid);
-        assert!(!transaction.is_checkpoint());
-
-        let checkpoint = TransactionFilePath::new_checkpoint(0, uuid);
-        assert!(checkpoint.is_checkpoint());
-    }
-
-    #[test]
-    fn test_transaction_file_path_deserialization() {
-        // Error cases
-        use TransactionFilePathParseError::*;
-
-        let mut df = DirsAndFileName::default();
-        let result = TransactionFilePath::from_relative_dirs_and_file_name(&df);
-        assert!(
-            matches!(result, Err(MissingRevisionCounter)),
-            "got {:?}",
-            result
-        );
-
-        df.push_dir("foo");
-        let result = TransactionFilePath::from_relative_dirs_and_file_name(&df);
-        assert!(
-            matches!(result, Err(InvalidRevisionCounter { .. })),
-            "got {:?}",
-            result
-        );
-
-        let mut df = DirsAndFileName::default();
-        df.push_dir("00000000000000000123");
-        df.push_dir("foo");
-        let result = TransactionFilePath::from_relative_dirs_and_file_name(&df);
-        assert!(
-            matches!(result, Err(UnexpectedDirectory)),
-            "got {:?}",
-            result
-        );
-
-        let mut df = DirsAndFileName::default();
-        df.push_dir("00000000000000000123");
-        let result = TransactionFilePath::from_relative_dirs_and_file_name(&df);
-        assert!(matches!(result, Err(MissingFileName)), "got {:?}", result);
-
-        df.set_file_name("foo");
-        let result = TransactionFilePath::from_relative_dirs_and_file_name(&df);
-        assert!(
-            matches!(result, Err(InvalidUuid { .. })),
-            "got {:?}",
-            result
-        );
-
-        let uuid = Uuid::new_v4();
-
-        df.set_file_name(&format!("{}", uuid));
-        let result = TransactionFilePath::from_relative_dirs_and_file_name(&df);
-        assert!(matches!(result, Err(MissingSuffix)), "got {:?}", result);
-
-        df.set_file_name(&format!("{}.exe", uuid));
-        let result = TransactionFilePath::from_relative_dirs_and_file_name(&df);
-        assert!(
-            matches!(result, Err(InvalidSuffix { .. })),
-            "got {:?}",
-            result
-        );
-
-        df.set_file_name(&format!("{}.{}.foo", uuid, TRANSACTION_FILE_SUFFIX));
-        let result = TransactionFilePath::from_relative_dirs_and_file_name(&df);
-        assert!(
-            matches!(result, Err(UnexpectedExtension)),
-            "got {:?}",
-            result
-        );
-
-        // Success case
-        df.set_file_name(&format!("{}.{}", uuid, TRANSACTION_FILE_SUFFIX));
-        let result = TransactionFilePath::from_relative_dirs_and_file_name(&df).unwrap();
-        assert_eq!(
-            result,
-            TransactionFilePath {
-                revision_counter: 123,
-                uuid,
-                suffix: TransactionFileSuffix::Transaction,
-            }
-        );
-        let round_trip = result.relative_dirs_and_file_name();
-        assert_eq!(round_trip, df);
-    }
-
-    #[test]
-    fn transaction_file_from_absolute() {
-        let object_store = make_object_store();
-
-        // Success case
-        let uuid = Uuid::new_v4();
-        let mut path = object_store.new_path();
-        path.push_all_dirs(&["dbs", "uuid", "data", "00000000000000000123"]);
-        path.set_file_name(&format!("{}.{}", uuid, CHECKPOINT_FILE_SUFFIX));
-        let result = TransactionFilePath::from_absolute(path);
-        assert_eq!(
-            result.unwrap(),
-            TransactionFilePath {
-                revision_counter: 123,
-                uuid,
-                suffix: TransactionFileSuffix::Checkpoint,
-            }
-        );
-
-        // Error cases
-        use TransactionFilePathParseError::*;
-
-        let mut path = object_store.new_path();
-        // incorrect directories are fine, we're assuming that list(transactions_path) scoped to the
-        // right directories so we don't check again on the way out
-        path.push_all_dirs(&["foo", "bar", "baz", "}*", "aoeu", "blah"]);
-        path.set_file_name("rules.pb");
-        assert_error!(
-            TransactionFilePath::from_absolute(path),
-            InvalidRevisionCounter { .. },
-        );
-
-        let mut path = object_store.new_path();
-        path.push_all_dirs(&["dbs", "uuid", "data", "00000000000000000123"]);
-        // missing file name
-        assert_error!(TransactionFilePath::from_absolute(path), MissingFileName,);
-    }
-
-    #[test]
-    fn transaction_file_relative_dirs_and_file_path() {
-        let uuid = Uuid::new_v4();
-        let tfp = TransactionFilePath {
-            revision_counter: 555,
-            uuid,
-            suffix: TransactionFileSuffix::Transaction,
-        };
-        let dirs_and_file_name = tfp.relative_dirs_and_file_name();
-        assert_eq!(
-            dirs_and_file_name.to_string(),
-            format!("00000000000000000555/{}.{}", uuid, TRANSACTION_FILE_SUFFIX)
-        );
-        let round_trip =
-            TransactionFilePath::from_relative_dirs_and_file_name(&dirs_and_file_name).unwrap();
-        assert_eq!(tfp, round_trip);
-    }
-
-    #[test]
-    fn transactions_path_join_with_parquet_file_path() {
-        let db_uuid = Uuid::new_v4();
-        let object_store = make_object_store();
-        let root_path = RootPath::new(&*object_store, db_uuid);
-        let iox_object_store = IoxObjectStore::existing(Arc::clone(&object_store), root_path);
-
-        let uuid = Uuid::new_v4();
-        let tfp = TransactionFilePath {
-            revision_counter: 555,
-            uuid,
-            suffix: TransactionFileSuffix::Checkpoint,
-        };
-
-        let path = iox_object_store.transactions_path.join(&tfp);
-
-        let mut expected_path = object_store.new_path();
-        expected_path.push_all_dirs(&[
-            ALL_DATABASES_DIRECTORY,
-            &db_uuid.to_string(),
-            "transactions",
-            "00000000000000000555",
-        ]);
-        expected_path.set_file_name(&format!("{}.{}", uuid, CHECKPOINT_FILE_SUFFIX));
-
-        assert_eq!(path, expected_path);
-    }
-}
