+++ iox_object_store_service/src/lib.rs
@@ -16,10 +16,10 @@ use futures::stream::BoxStream;
 use futures::StreamExt;
 use generated_types::influxdata::iox::object_store::v1::*;
 use iox_catalog::interface::Catalog;
-use iox_object_store::ParquetFilePath;
 use object_store::DynObjectStore;
 use observability_deps::tracing::*;
-use std::sync::Arc;
+use parquet_file::ParquetFilePath;
+use std::{ops::Deref, sync::Arc};
 use tonic::{Request, Response, Status};
 use uuid::Uuid;
 
@@ -66,15 +66,14 @@ impl object_store_service_server::ObjectStoreService for ObjectStoreService {
             })?
             .ok_or_else(|| Status::not_found(req.uuid))?;
 
-        let path = ParquetFilePath::new_new_gen(
+        let path = ParquetFilePath::new(
             parquet_file.namespace_id,
             parquet_file.table_id,
             parquet_file.sequencer_id,
             parquet_file.partition_id,
             parquet_file.object_store_id,
-        )
-        .absolute_dirs_and_file_name();
-        let path = self.object_store.path_from_dirs_and_filename(path);
+        );
+        let path = path.object_store_path(self.object_store.deref());
 
         let res = self
             .object_store
@@ -164,15 +163,14 @@ mod tests {
 
         let object_store = Arc::new(ObjectStoreImpl::new_in_memory());
 
-        let path = ParquetFilePath::new_new_gen(
+        let path = ParquetFilePath::new(
             p1.namespace_id,
             p1.table_id,
             p1.sequencer_id,
             p1.partition_id,
             p1.object_store_id,
-        )
-        .absolute_dirs_and_file_name();
-        let path = object_store.path_from_dirs_and_filename(path);
+        );
+        let path = path.object_store_path(object_store.deref());
 
         let data = Bytes::from_static(b"some data");
