+++ iox_tests/Cargo.toml
@@ -11,7 +11,6 @@ bytes = "1.0"
 data_types2 = { path = "../data_types2" }
 datafusion = { path = "../datafusion" }
 iox_catalog = { path = "../iox_catalog" }
-iox_object_store = { path = "../iox_object_store" }
 metric = { path = "../metric" }
 mutable_batch_lp = { path = "../mutable_batch_lp" }
 object_store = { path = "../object_store" }
