+++ querier/Cargo.toml
@@ -19,7 +19,6 @@ generated_types = { path = "../generated_types" }
 hyper = "0.14"
 influxdb_iox_client = { path = "../influxdb_iox_client" }
 iox_catalog = { path = "../iox_catalog" }
-iox_object_store = { path = "../iox_object_store" }
 metric = { path = "../metric" }
 object_store = { path = "../object_store" }
 observability_deps = { path = "../observability_deps" }
